Upgrade to Rails 4.2.7.1
There is no patch for the 4.1.x version of Rails to address CVE-2016-6316 and CVE-2016-6317 so we have to upgrade to 4.2.x first.

I've adopted the approach documented at http://guides.rubyonrails.org/upgrading_ruby_on_rails.html which introduces new files and changes of old files in accordance with the Rails application generator.